pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "sandboxer"
version = "0.1.0"
description = "Docker Sandbox manager for autonomous agents — templates, profiles, and unified UI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
sandboxer = "sandboxer.cli:entrypoint"

[tool.pytest.ini_options]
markers = [
    "integration: tests that require Docker Desktop with sandbox support",
]

[tool.hatch.build.targets.wheel]
packages = ["sandboxer"]
